requirements.txt
google-generativeai==0.3.1
python-dotenv==1.0.0
pillow==11.1.0
pytest==7.4.0
pytest-cov==4.1.0 